requirements (2).txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-mistralai>=0.1.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
python-dotenv>=1.0.0
